src/main/webapp/app/entities/view/route/view-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IView, View } from '../view.model';
import { ViewService } from '../service/view.service';

import { ViewRoutingResolveService } from './view-routing-resolve.service';

describe('Service Tests', () => {
  describe('View routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: ViewRoutingResolveService;
    let service: ViewService;
    let resultView: IView | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(ViewRoutingResolveService);
      service = TestBed.inject(ViewService);
      resultView = undefined;
    });

    describe('resolve', () => {
      it('should return IView returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultView = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultView).toEqual({ id: 123 });
      });

      it('should return new IView if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultView = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultView).toEqual(new View());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as View })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultView = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultView).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
